@@ -1,54 +1,20 @@
-﻿using System;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata;
+﻿using Microsoft.EntityFrameworkCore;
 
 namespace WorkOrderApp.Shared.Models
 {
-    public partial class OrderDataContext : DbContext
+    public class OrderDataContext : DbContext
     {
         public OrderDataContext()
-        {
-        }
-
+        { }
+        
         public OrderDataContext(DbContextOptions<OrderDataContext> options)
             : base(options)
         {
         }
 
-        public virtual DbSet<WorkOrder> WorkOrder { get; set; }
-
         protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            if (!optionsBuilder.IsConfigured)
-            {
-#warning To protect potentially sensitive information in your connection string, you should move it out of source code. See http://go.microsoft.com/fwlink/?LinkId=723263 for guidance on storing connection strings.
-                optionsBuilder.UseSqlServer("Data Source=(LocalDB)\\MSSQLLocalDB;AttachDbFilename=D:\\WorkOrderApp-286866234\\WorkOrderApp\\WorkOrderApp\\Shared\\App_Data\\NORTHWND.MDF;Integrated Security=True");
-            }
-        }
+            => optionsBuilder.UseSqlite("Data Source=database.db");
 
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            modelBuilder.HasAnnotation("ProductVersion", "2.2.0-rtm-35687");
-
-            modelBuilder.Entity<WorkOrder>(entity =>
-            {
-                entity.HasKey(e => e.OrderId)
-                    .HasName("PK__WorkOrde__C3905BAFEBEF45D7");
-
-                entity.Property(e => e.OrderId)
-                    .HasColumnName("OrderID")
-                    .ValueGeneratedNever();
-
-                entity.Property(e => e.Description).HasMaxLength(50);
-
-                entity.Property(e => e.Location).HasMaxLength(50);
-
-                entity.Property(e => e.OrderDate).HasColumnType("date");
-
-                entity.Property(e => e.Status).HasMaxLength(50);
-
-                entity.Property(e => e.Title).HasMaxLength(10);
-            });
-        }
+        public DbSet<WorkOrder> WorkOrder { get; set; }
     }
 }
